property: Add support to fwnode_graph_get_endpoint_by_id() for parsing fwnode->secondary if no endpoint is found under the primary
fwnode_handles created from software_nodes can be used to construct
media graphs for devices where the firmware is lacking. Implementation
of this method requires fwnode_graph_get_endpoint_by_id() to also check
fwnode->secondary in the event that it fails to find a valid endpoint
against the primary.